requirements.txt
flask
line-bot-sdk
requests
uwsgi
pymysql
bs4